// src/utility/types/admin/policy/policyType.ts
// all policy api response 


export interface AllPolicyApiResponse {
    id: number;
    name: string;
    slug: string;
    description: string;
    logo_url: string;
    status: string;
    created_at: string;
    updated_at: string;
}


// single policy api response 





export interface SinglePolicyApiResponseType {
    id: number;
    name: string;
    slug: string;
    description: string;
    logo_url: string;
    status: string;
    created_at: string;
    updated_at: string
}